Mms: Need to hide keyboard when the Popup loses focus
Previous commit, prevented QuickMessage text from disappearing, but was leaving
the keyboard open, so user could continue to type. Unfortunately, the EditText
loses focus (not sure if that can be changed?), so the typed text goes nowhere
and causes confusion.

This commit will hide the keyboard, so user will need to reselect the EditText
field first to redisplay the keyboard and continue typing.

Change-Id: I4869392b16be015ab9a8cd99bf3f95f0e86b74ad